feat(system):  improving layout and usability adds masks on values types - #21 (#22)
* feat(system): implemented maskMoney, remove plugin mask alpineJS

* feat(system): implemented cast moneyToDatabase;
refactor(transaction): save using fill,
feat(transaction): included setValueAttribute;

* refactor(system): improve maskMoney.
feat(transaction): added mask on value

* feat(transaction): index, improve layout

* refactor(system): modify type columns values

* fix(payment type):solved not is default

+++ database/migrations/2022_08_26_003153_create_contracts_table.php
@@ -18,8 +18,8 @@ class CreateContractsTable extends Migration
             $table->foreignId('user_id')->constrained();
             // $table->foreignId('person_id')->constrained();
             $table->string('title');
-            $table->double('value');
-            $table->double('prediction');
+            $table->decimal('value', 20, 2);
+            $table->decimal('prediction', 20, 2);
             $table->timestamp('date_init');
             $table->timestamp('date_end')->nullable();
             $table->enum('status', ['active','inactive'])->default('active');
